ssss
ssss

+++ Initialization.c
@@ -3,7 +3,7 @@
 //
 // Created by Artem on 28.02.2023.
 //
-Computer init(char name[100], float price, int purpose, int operationSystem, char CPU[50], char GPU[50]) {
+Computer init(char name[100], int price, int purpose, int operationSystem, char CPU[50], char GPU[50]) {
     Computer computer;
     snprintf(computer.name, sizeof(computer.name), "%s", name);
     computer.price = price;
